pyproject.toml
[project]
name = "age-detection-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.11",
    "httpx>=0.28.1",
    "pillow>=11.1.0",
    "pre-commit>=4.1.0",
    "pydantic>=2.10.6",
    "pytest>=8.3.5",
    "python-multipart>=0.0.20",
    "ruff>=0.9.10",
    "sentencepiece>=0.2.1",
    "torch>=2.6.0",
    "transformers>=4.49.0",
    "uvicorn>=0.34.0",
]
